+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2024-10-06
+### Added
+- new Feature to show analysis graphs on overview page.
+- New graph to show total transaction amount vs total tax amount.
+- New graph to show category wise sales amount over time.
+- New graph to show total number of transaction over time.
+- New graph to show total transaction amount over time.
+- New graph to show inventory quantity and select based on category.
+- Implemented APIs to the graphs.
+
+### Changes
+- CSS changes to show the proper visibility on header and footer.
+
+**NOTE:** Versions progressed from 0.2.1 to 0.2.9, incorporating multiple new features into the application.
+
 ## [0.3.0] - 2024-10-06
 ### Added
 - Added new Feature to list transactions.